fix issue of docker client not able to write cmd to tmp file when dir got deleted after NM process started
Summary: fix issue of docker client not able to write cmd to tmp file when dir got deleted after NM process started

Reviewers: pgolash, O2123 User sunchao: Add reviewers, O2125 User ekanth: Add reviewers, O2127 User pgolash: Add reviewers, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers!

Subscribers: O2131 Project Hadoop-HDFS Dev: Add subscribers, O2133 Project Data Foundation - Hadoop Ecosystem: Add subscribers

Maniphest Tasks: T6155353

Differential Revision: https://code.uberinternal.com/D4569061

+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-nodemanager/src/main/java/org/apache/hadoop/yarn/server/nodemanager/containermanager/linux/runtime/docker/DockerClient.java
@@ -66,9 +66,16 @@ public final class DockerClient {
       throws ContainerExecutionException {
     File dockerCommandFile = null;
     try {
+      File tmpDir = new File(tmpDirPath);
+      // if directory of conf(hadoop.tmp.dir) get deleted during nm process is running, re-created it before
+      // write tmp file under the dir.
+      if (!(tmpDir.exists() || tmpDir.mkdirs())) {
+        LOG.warn("Unable to create directory when write command to temp file: " + tmpDirPath);
+        throw new ContainerExecutionException("Unable to create directory: " +
+            tmpDirPath);
+      }
       dockerCommandFile = File.createTempFile(TMP_FILE_PREFIX + filePrefix,
-          TMP_FILE_SUFFIX, new
-          File(tmpDirPath));
+          TMP_FILE_SUFFIX, tmpDir);
 
       Writer writer = new OutputStreamWriter(
           new FileOutputStream(dockerCommandFile), "UTF-8");
